Cleaned Up models folder
